fix: update `Balance`
- changes the Amount type to float64

+++ internal/application/usecase/get_account.go
@@ -46,6 +46,6 @@ type GetAccountOutput struct {
 }
 
 type getAccountOutpuBalance struct {
-	AssetID string `json:"asset_id"`
-	Amount  uint64 `json:"amount"`
+	AssetID string  `json:"asset_id"`
+	Amount  float64 `json:"amount"`
 }
